Update internal architecture
Too many moving parts create a big commit. Summary:

- Left and right versions of the conflicted model are loaded as internal
resources. These will be used later to get the details to present the
conflicts, to apply conflict resolution actions, and to save the model
(including partial saving with models that still have conflicts)

- Conflicts are now stored in the resource, where they will be read by
the editor that comes next. We also store conflict sections with those
elements that are not part of a conflict (but are new and coming from a
left or right conflict section), for information purposes.

- Everything is adapted to the new Conflicts hierarchy, and concrete
details for each conflict can be retrieved from the left and right
resources thanks to the XMI ids.