Add initial conflicts editor
This is a very preliminar version, where the joined model that is parsed
while processing is depicted. This model has:

- Left and right elements in conflict sections that are not conflicting
- Left and right elements in conflicts, where possible
- Only right elements when not possible (e.g. in containment 1-bounded
references, and in terms of attribute values)
- Any common element

Also, as the two internal left and right resources are loaded in the
same ResourceSet, they are shown in the default visualisation of the
Ecore editor, which is something I'll need to tweak in the future.