Load left and right resources based on conflict lines
A new resource implementation (ConflictVersionResource) is required, as
the lines locator that the xmi handlers can access is "situational"
(e.g. tells the element line in methods such as startElement,
endElement, etc)

Left and right resources are loaded, and ConflictSections are identified
and loaded with any element detected inside them (based on the line
types identified by the ConflictsPreprocessor)

Next: process the ConflictSections to identify actual conflicts.

Also: clean PeaceMakerXMIHandler from any processing methods (the resource is not "loaded" as typically, maybe I need to check if that can bring problems later (e.g. when saving it))