Add initial listeners to radio actions, including undo and save
It's a bit clunky, as some cases are not supported, but now it is
possible to select resolve actions (once per conflict), undo previous
selections (this does not update the radio buttons as of yet), and save
the resource with the resolved actions.

TODO: apart from the above, I might need to save left and right names of the conflict sections separators (e.g. HEAD or branch names, commit tags, whichever is used by the diff conflict creator), although this might not be necessary in the end.