Move conflict initialisations to specific init method
The PeaceMakerResource was responsible to set any additional information
that a conflict might need to work when adding it to the list (after
identification).

Now that initialisation has been extracted into a more natural place,
i.e., a method in each specific conflict class receiving the resource
(e.g. to gather left and right version objects for copying or removing
stuff)