Add "show in viewers" link to the conflicts
Now each conflict has a link that allows to select the relevant object
in the left and right version resources. Also, conflicts now always have
access to the peacemaker resource (instead of the previous init method),
which allows for getting the current value of the left and right
resources. This makes the link event work also when some actions have
redefined the original objects.