Add a replace method that prevents containment and reference issues
When replacing objects between left and right resources, the previous
solution (copyAndReplace) generated a copy of the replacing object,
which was swapped with the replaced one. This solution creates issues:

- all contained children (which could have their own conflicts) would
also be forcefully replaced
- any reference pointing to the replaced element would be dangling after
the swap (problems when saving)

The new solution just copies the internals of the replacing element into
the replaced one. The scope of the internals contains:

- The xmi:id of the object
- Any EAttribute
- Non-containment EReferences. These are also properly treated to point
to objects in the resource of the replaced object (i.e. the external
references problem we solve in other cases by calling
fixExternalReferences)

As of now, cointained children would be treated as independent objects
if they are part of a conflict. If they are not, my current guess is
that the children on both sides are the same, as the version control
merge algorithm would have combined both parts into the COMMON section
of the conflicted file.