Reorder metamodel contents to avoid forward references
The inclusion of references in psl models (task efforts pointing to the
responsible person) highlighted a problem in the features order: as
people were the last feature of the project, all created efforts
(thousands) needed to be postprocessed at the end of the document load,
because of the reference to a person that was not loaded yet. So,
performance measurements were completely obscured by these forward
references resolution. Not any more, I hope.