Fix save bug when ancestor is not loaded to identify conflicts
The ancestor version of a model is not loaded if not needed for the
proper identification of conflicts. However, if present, it must be
loaded when saving the file to prevent the loss of ancestor information
if there are conflicts still present in the file.

Maybe this could also be avoided if we know the model we are saving has
all conflicts resolved.