Add KeepDelete conflict type
Sometimes objects in a conflict section have been deleted in one side,
but not modified (i.e. just kept) in the other (but are part of conflict
sections nonetheless). For some cases, deleting these objects could be
problematic, so this conflict type seeks to leave that decision to the
final user.