Use custom XMI handler for better finding dup ids
Using the pool of handlers when loading, now we provide a custom
PeacemakerXMIHandler instead of decorating the one used by default from
the resource.

This change allows a better management of duplicated ids: now we can
access the object stack (not possible in the handler decorator) to
gather elements so that we can link to them in the editor, or we could
even allow removing unwanted copies.

If a concrete resource has handler modifications that are needed for
loading the model, it should be possible to extend such a handler with
the (very minor) functionality of our custom handler, and provide that
version in the pool of handlers. It's something that could be managed
with an extension point.