Add vertical scrollbar to the conflicts list
A bit buggy still, for instance the scroll bar is not shown when first
opening the editor (if necessary) until resizing the editor window.

Also, events on the expandable conflicts get the previous conflict lists
size, so it is not possible to update the scrolled composite to the new
minHeight.